requirements.txt
python-dotenv
openai
pinecone-client
langchain
pypdf
tiktoken
bs4
